more dd tests
